Refactor utility functions
To make the windows and unix separation cleaner.
